requirements.txt
aiohttp==3.8.5
asyncio==3.4.3
pandas==1.5.3
python-dotenv==0.21.1
requests==2.28.2
XlsxWriter==3.0.8
